Update useCachedResources code block example in RN
Replacing the function in the way the doc shows throws an error for the Ionicons import. This version works on the template the doc intends to be used if you spin it up with the latest expo sdk.

The documentation wording makes it sound like you're supposed to replace the whole file with that function anyway this will be more clear to someone who may not understand the custom hook implementation.